multiverse.py
```python
from universe import Universe

class Multiverse:
    def __init__(self):
        """
        Initialize a new Multiverse instance.
        """
        self.universes = {}

    def add_universe(self, id, properties):
        """
        Add a new universe to the multiverse.

        :param id: The unique identifier for the new universe.
        :param properties: A dictionary containing the properties of the new universe.
        """
        self.universes[id] = Universe(id, properties)

    def remove_universe(self, id):
        """
        Remove a universe from the multiverse.

        :param id: The unique identifier for the universe to remove.
        """
        if id in self.universes:
            del self.universes[id]

    def get_universe(self, id):
        """
        Get a universe from the multiverse.

        :param id: The unique identifier for the universe to get.
        :return: The Universe instance with the given id, or None if no such universe exists.
        """
        return self.universes.get(id)

    def simulate(self):
        """
        Simulate all universes in the multiverse. This method should call the simulate method
        on each Universe instance in the multiverse.
        """
        for universe in self.universes.values():
            universe.simulate()

    def get_properties(self, id):
        """
        Get the properties of a universe in the multiverse.

        :param id: The unique identifier for the universe whose properties to get.
        :return: A dictionary containing the properties of the universe, or None if no such universe exists.
        """
        universe = self.get_universe(id)
        return universe.get_properties() if universe else None

    def set_properties(self, id, properties):
        """
        Set the properties of a universe in the multiverse.

        :param id: The unique identifier for the universe whose properties to set.
        :param properties: A dictionary containing the new properties for the universe.
        """
        universe = self.get_universe(id)
        if universe:
            universe.set_properties(properties)
```
