tests/test_multiverse.py
```python
import unittest
from multiverse import Multiverse

class TestMultiverse(unittest.TestCase):
    def setUp(self):
        """
        Set up a new Multiverse instance before each test.
        """
        self.multiverse = Multiverse()

    def test_add_universe(self):
        """
        Test that a universe can be added to the multiverse.
        """
        properties = {"property1": "value1", "property2": "value2"}
        self.multiverse.add_universe("universe1", properties)
        self.assertIn("universe1", self.multiverse.universes)

    def test_remove_universe(self):
        """
        Test that a universe can be removed from the multiverse.
        """
        properties = {"property1": "value1", "property2": "value2"}
        self.multiverse.add_universe("universe1", properties)
        self.multiverse.remove_universe("universe1")
        self.assertNotIn("universe1", self.multiverse.universes)

    def test_get_universe(self):
        """
        Test that a universe can be retrieved from the multiverse.
        """
        properties = {"property1": "value1", "property2": "value2"}
        self.multiverse.add_universe("universe1", properties)
        universe = self.multiverse.get_universe("universe1")
        self.assertEqual(universe.get_properties(), properties)

    def test_simulate(self):
        """
        Test that the simulate method works correctly.
        """
        properties = {"property1": "value1", "property2": "value2"}
        self.multiverse.add_universe("universe1", properties)
        self.multiverse.simulate()
        # TODO: Add assertions to check the state of the universe after simulation

    def test_get_properties(self):
        """
        Test that the properties of a universe can be retrieved.
        """
        properties = {"property1": "value1", "property2": "value2"}
        self.multiverse.add_universe("universe1", properties)
        retrieved_properties = self.multiverse.get_properties("universe1")
        self.assertEqual(retrieved_properties, properties)

    def test_set_properties(self):
        """
        Test that the properties of a universe can be set.
        """
        properties = {"property1": "value1", "property2": "value2"}
        self.multiverse.add_universe("universe1", properties)
        new_properties = {"property1": "new_value1", "property2": "new_value2"}
        self.multiverse.set_properties("universe1", new_properties)
        retrieved_properties = self.multiverse.get_properties("universe1")
        self.assertEqual(retrieved_properties, new_properties)

if __name__ == "__main__":
    unittest.main()
```
